agregado imagenes en vista publicaciones
