+++ src/app/pages/home/home.page.ts
@@ -3,10 +3,6 @@ import { Component, OnInit } from '@angular/core';
 import { PublicacionI } from "../../models/publicacion.interface";
 import { PublicacionService } from "../../services/publicacion.service";
 
-import { AngularFireStorage } from "angularfire2/storage";
-import { Observable } from "rxjs";
-import { finalize } from 'rxjs/operators';
-
 
 @Component({
   selector: 'app-home',
@@ -20,12 +16,7 @@ export class HomePage implements OnInit {
   public goalList :any [];
   public loadedGoalList : any [];
 
-  // propiedades de storage
-  uploadPercent: Observable<number>;
-  downloadURL: Observable<string>;
-  image: string = null;
-
-  constructor(private publicacionesService: PublicacionService, private storage: AngularFireStorage) {
+  constructor(private publicacionesService: PublicacionService) {
   }
 
   ngOnInit(){
@@ -56,25 +47,6 @@ export class HomePage implements OnInit {
         return false;
       }
     });
-  }
-
-  // metodos de storage
-  subirArchivo(event) {
-    const file = event.target.files[0];
-    const path = "images/demo126";
-    const task = this.storage.upload(path, file);
-    const ref = this.storage.ref(path);
-
-    // observar porcentaje cambiados
-    this.uploadPercent = task.percentageChanges();
-    console.log('Image uploaded!');
-    task.snapshotChanges().pipe(
-      finalize(() => {
-        this.downloadURL = ref.getDownloadURL()
-        this.downloadURL.subscribe(url => (this.image = url));
-      })
-    ).subscribe();
-
   }
