+++ src/app/pages/publicacion/publicacion.page.html
@@ -19,8 +19,19 @@
       <ion-input required type="text" placeholder="descripcion" [(ngModel)]="publicacion.descripcion"></ion-input>
     </ion-item>
 
+    <input type="file" (change)="subirArchivo($event)">
+    <div>{{ uploadPercent | async }}</div>
+    <!-- <a [href]="downloadURL | async">{{ downloadURL | async }}</a> -->
+
+    <ion-card>
+      <ion-card-header>
+        <img [src]="downloadURL | async" alt="" srcset="">
+        <ion-card-subtitle>{{ downloadURL | async }}</ion-card-subtitle>
+      </ion-card-header>
+    </ion-card>
+
     <ion-item *ngIf="publicacion">
-      <ion-input required type="text" placeholder="imagen" [(ngModel)]="publicacion.imagen"></ion-input>
+      <ion-input required type="text" placeholder="imagen" [(ngModel)]="publicacion.imagen" value="{{ downloadURL | async }}"></ion-input>
     </ion-item>
 
   </ion-list>
